the update code 20-12

+++ be_booking_hotel/e_be_booking_hotel/models/hotel.py
@@ -1,5 +1,6 @@
 from django.db import models
 
+
 class Hotel(models.Model):
     hotel_id = models.AutoField(primary_key=True)
     hotel_name = models.CharField(max_length=255)
@@ -7,8 +8,10 @@ class Hotel(models.Model):
     address = models.CharField(max_length=255)
     city = models.CharField(max_length=100)
     description = models.TextField()
-    rating = models.DecimalField(max_digits=3, decimal_places=1, null=True, blank=True)
+    rating = models.FloatField(null=True, blank=True)  # Thay đổi kiểu dữ liệu
+    # rating = models.DecimalField(max_digits=3, decimal_places=1, null=True, blank=True)
     contact_info = models.TextField()
 
     def __str__(self):
         return self.hotel_name
+
